feat: use str.contains instead of str.match
Use str.contains to search for a match. This allows for the pattern to appear at any point in the string requiring the patterns to be less specific. This makes sense because there may often be matches of a single string at any point in the input needed due to the nature of the reports.