fix: allow for unknown Enum for Column functions
Because Enum do not allow for inheritance in Python 3.12.1, each Column enum is defined separately in its own files. This means to type check Column types, it would have to be cross verified with each Column type to be one of those (DataSet.Column, AppointmentDataSet.Column, EnrollmentDataSet.Column, ReferralDataSet.Column, SurveyDataSet.Column). If new dataset types are added, they will have to be defined in every location, which is not ideal. Use a slightly more risky Enum type check.